refactor: refactored user activities architecture
