+++ src/app/app.module.ts
@@ -3,7 +3,7 @@ import { DatabaseModule } from '../database/database.module';
 import { UserModule } from '../modules/user/infra/modules/user.module';
 import { UserScoreModule } from '../modules/user-score/infra/modules/user-score.module';
 import { CourseModule } from '../modules/course/infra/modules/course.module';
-import { UserActivitiesAnsweredModule } from '../modules/user-activities-answered/user-activities-answered.module';
+import { UserActivitiesAnsweredModule } from '../modules/user-activities-answered/infra/modules/user-activities-answered.module';
 import { UserCoursesConcludedModule } from '../modules/user-courses-concluded/user-courses-concluded.module';
 import { ActivityModule } from '../modules/activity/infra/modules/activity.module';
