+++ src/modules/user-activities-answered/domain/dtos/errors/ActivityAlreadyAnswered.exception.ts
