+++ src/modules/user-activities-answered/domain/dtos/errors/ActivityNotFound.exception.ts
