+++ src/modules/user-activities-answered/domain/dtos/errors/WrongAnswer.exception.ts
