+++ src/modules/user-activities-answered/domain/dtos/repositories/user-activities-answered.ts
@@ -1,0 +1,5 @@
+import { UserActivityAnswered } from "../../../infra/db/entity/user-activities-answered.entity";
+
+export interface UserActivitiesAnsweredRepositoryInterface {
+    findAnsweredActivity(user_id: number, activity_id: number): Promise<UserActivityAnswered | null>;
+}
