--- src/modules/user-activities-answered/dto/user-activities-answered.dto.spec.ts
@@ -1,7 +1,0 @@
-import { UserActivitiesAnsweredDto } from './user-activities-answered.dto';
-
-describe('UserActivitiesAnsweredDto', () => {
-  it('should be defined', () => {
-    expect(new UserActivitiesAnsweredDto()).toBeDefined();
-  });
-});
