--- src/modules/user-activities-answered/dto/user-activities-answered.dto.ts
@@ -1,1 +1,0 @@
-export class UserActivitiesAnsweredDto {}
