+++ src/modules/user-activities-answered/infra/db/entity/user-activities-answered.entity.spec.ts
