+++ src/modules/user-activities-answered/infra/db/repositories/user-activities-answered.repository.spec.ts
