+++ src/modules/user-activities-answered/infra/db/repositories/user-activities-answered.repository.ts
@@ -1,9 +1,10 @@
 import { Injectable } from '@nestjs/common';
 import { DataSource, Repository } from 'typeorm';
 import { UserActivityAnswered } from '../entity/user-activities-answered.entity';
+import { UserActivitiesAnsweredRepositoryInterface } from '../../../domain/dtos/repositories/user-activities-answered';
 
 @Injectable()
-export class UserActivityAnsweredRepository extends Repository<UserActivityAnswered> {
+export class UserActivityAnsweredRepository extends Repository<UserActivityAnswered> implements UserActivitiesAnsweredRepositoryInterface {
   constructor(private dataSource: DataSource) {
     super(UserActivityAnswered, dataSource.createEntityManager());
   }
