+++ src/modules/user-activities-answered/infra/services/user-activities-answered.service.ts
@@ -1,13 +1,13 @@
 import { Injectable } from '@nestjs/common';
 import { ActivityRepository } from '../../activity/repositories/activity.repository';
-import { UserActivityAnsweredRepository } from '../repository/user-activities-answered.repository';
+import { UserActivityAnsweredRepository } from '../repositories/user-activities-answered.repository';
 import { UserRepository } from '../../user/repositories/user.repository';
-import { UserNotFoundException } from '../../user/domain/dtos/errors/UserNotFound.exception';
-import { WrongAnswerException } from '../domain/errors/WrongAnswer.exception';
-import { ActivityNotFoundException } from '../domain/errors/ActivityNotFound.exception';
-import { ActivityAlreadyAnsweredException } from '../domain/errors/ActivityAlreadyAnswered.exception';
-import { UserCourseConcludedRepository } from '../../user-courses-concluded/repository/user-courses-concluded.repository';
-import { UserScoreService } from '../../user-score/infra/services/user-score.service';
+import { UserNotFoundException } from '../../../user/domain/dtos/errors/UserNotFound.exception';
+import { WrongAnswerException } from '../../domain/dtos/errors/WrongAnswer.exception';
+import { ActivityNotFoundException } from '../../domain/dtos/errors/ActivityNotFound.exception';
+import { ActivityAlreadyAnsweredException } from '../../domain/dtos/errors/ActivityAlreadyAnswered.exception';
+import { UserCourseConcludedRepository } from '../../../user-courses-concluded/repository/user-courses-concluded.repository';
+import { UserScoreService } from '../../../user-score/infra/services/user-score.service';
 import { CourseRepository } from '../../course/repositories/course.repository';
 
 @Injectable()
